pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oil-forecast-project"
version = "0.1.0"
description = "Reproducible evaluation of long-term oil price forecasts for the UCB energy project."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "beautifulsoup4>=4.12",
  "matplotlib>=3.8",
  "numpy>=1.24",
  "openpyxl>=3.1",
  "python-docx>=1.1",
  "pandas>=2.0",
  "python-pptx>=1.0",
  "requests>=2.31",
  "scikit-learn>=1.3",
  "seaborn>=0.13",
  "statsmodels>=0.14",
  "XlsxWriter>=3.1",
  "yfinance>=0.2.54",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
